% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getheart}
\alias{getheart}
\title{Get the full data set of the ECG recordings in micro-volt (μV).}
\usage{
getheart(token, signalid)
}
\arguments{
\item{token}{Your token obtained using \code{withings_auth()}}

\item{signalid}{ID of a signal. Use \code{getheartlist()} to get one.}
}
\value{
A list containing the status (status) and the data (body)
}
\description{
Get the full data set of the ECG recordings in micro-volt (μV).
}
\examples{
\dontrun{
getheart(token, 123456)
}

}
