% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getmeas}
\alias{getmeas}
\title{Get measures data}
\usage{
getmeas(
  token,
  meastype,
  category,
  startdate = NULL,
  enddate = NULL,
  offset = NULL,
  lastupdate = NULL,
  tz = ""
)
}
\arguments{
\item{token}{Your token obtained using \code{withings_auth()}}

\item{meastype}{Which measure number should be obtained (see API)}

\item{category}{Which category number should be obtained (see API)}

\item{startdate}{Start date in character, date or POSIXct format}

\item{enddate}{End date in character, date or POSIXct format}

\item{offset}{Offset}

\item{lastupdate}{Last update (see API)}

\item{tz}{Time zone}
}
\value{
A list containing the status (status) and the data (body)
}
\description{
Get measures data
}
\examples{
\dontrun{
getmeas(token, meastype = 1, category=1, "2019-01-01", "2019-10-18")
}
}
