% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getsleepsummary}
\alias{getsleepsummary}
\title{Get sleep summary data}
\usage{
getsleepsummary(token, startdate = NULL, enddate = NULL, tz = "")
}
\arguments{
\item{token}{Your token obtained using \code{withings_auth()}}

\item{startdate}{Start date in character, date or POSIXct format}

\item{enddate}{End date in character, date or POSIXct format}

\item{tz}{Time zone}
}
\value{
A list containing the status (status) and the data (body)
}
\description{
Get sleep summary data
}
\examples{
\dontrun{
getsleepsummary(token, "2018-10-16", "2018-10-18")
}

}
