% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getworkouts}
\alias{getworkouts}
\title{Get workouts data}
\usage{
getworkouts(
  token,
  startdate = NULL,
  enddate = NULL,
  offset = NULL,
  lastupdate = NULL,
  tz = ""
)
}
\arguments{
\item{token}{Your token obtained using \code{withings_auth()}}

\item{startdate}{Start date in character, date or POSIXct format}

\item{enddate}{End date in character, date or POSIXct format}

\item{offset}{Offset}

\item{lastupdate}{Last update (see API)}

\item{tz}{Time zone}
}
\value{
A list containing the status (status) and the data (body)
}
\description{
Get workouts data
}
\examples{
\dontrun{
getworkouts(token, "2018-10-16", "2018-11-18")
}
}
