% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{read_csv_startdurval}
\alias{read_csv_startdurval}
\title{Read CSV including start duration value}
\usage{
read_csv_startdurval(filename, tz = Sys.timezone())
}
\arguments{
\item{filename}{CSV filename}

\item{tz}{Time zone}
}
\value{
A tibble including the munged values in tidy format
}
\description{
This function reads the raw input data files which can be downloaded by
requesting one's data from the Withings Health Mate site.
}
\examples{
\dontrun{
read_csv_startdurval('raw_tracker_hr.csv')
}

}
