% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{startdurval}
\alias{startdurval}
\title{Fix CSV files with start, duration and value}
\usage{
startdurval(start, dur, value)
}
\arguments{
\item{start}{String start value}

\item{dur}{String duration value}

\item{value}{String value value}
}
\value{
A tibble with each observation on a new line
}
\description{
Fix CSV files with start, duration and value
}
\examples{
\dontrun{
startdurval(Sys.time(),
            "29,60,60,60", "23,9,9,35" )
}
}
