% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{withings_auth}
\alias{withings_auth}
\title{Authenticate with the Withings API}
\usage{
withings_auth(client_id, client_secret, scope = c("user.metrics", "user.info",
  "user.activity"),
  authorize_url = "https://account.withings.com/oauth2_user/authorize2",
  access_url = "https://account.withings.com/oauth2/token")
}
\arguments{
\item{client_id}{Your client ID}

\item{client_secret}{Your Client Secret}

\item{scope}{Which information you would like to access}

\item{authorize_url}{The authorize URL (you shouldn't need to touch this)}

\item{access_url}{The access URL (you shouldn't need to touch this)}
}
\value{
An access token, or refreshing the access token.
}
\description{
This function performs the OAuth dance to get an OAuth2.0 token
}
\examples{
\dontrun{
client_id <- 12345
client_secret <- "hunter2"
token <- withings_auth(client_id, client_secret)
}
}
