% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{withings_authrefresh}
\alias{withings_authrefresh}
\title{Refresh token}
\usage{
withings_authrefresh(token)
}
\arguments{
\item{token}{Your token obtained using \code{withings_auth()}}
}
\value{
A refreshed access token
}
\description{
Refresh token
}
\examples{
\dontrun{
client_id <- 12345
client_secret <- "hunter2"
token <- withings_auth(client_id, client_secret)

token <- withings_authrefresh(token)
}
}
