% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getheart}
\alias{getheart}
\title{Get list of ECG recordings and Afib classification in the chosen period of
time}
\usage{
getheart(token, signalid)
}
\arguments{
\item{token}{Your token obtained using \code{withings_auth()}}

\item{signalid}{ID of a signal. Use \code{getheartlist()} to get one}
}
\value{
Get the full data set of the ECG recordings in micro-volt (μV).
}
\description{
Get list of ECG recordings and Afib classification in the chosen period of
time
}
\examples{
\dontrun{
getheart(token, 123456)
}

}
